requirements.txt
sentence-transformers
faiss-cpu
rank-bm25
pypdf
numpy
scikit-learn
fastapi
unicorn
python-dotenv